+++ CHANGELOG.md
@@ -33,6 +33,9 @@ Version template:
 ### Fixed
 
 * Sharded solr monitoring throws exception in Alfresco 6.2
+* Fixed and refactored code for solr, so that it works for ASS>=2.0.0. Improved handling of solrconfig.xml for solr6.
+
+[#74] https://github.com/xenit-eu/alfred-telemetry/pull/74
 
 ## [0.4.0] - 2020-01-12